cattle_breed_mvp/requirements.txt
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.1
tqdm==4.66.1
albumentations==1.3.1
timm==0.9.7
tensorboard==2.14.1
streamlit==1.28.0
